+++ test/python/unittest/API/APITest.py
@@ -23,7 +23,9 @@ except ImportError:
 
 from TestUtilities.TestUtilities import generateTestFilePath
 
-tested_file_extensions = [ext for ext in io.file_extensions if ext != 'sst']
+tested_file_extensions = [
+    ext for ext in io.file_extensions if ext != 'sst' and ext != 'ssc'
+]
 
 
 class APITest(unittest.TestCase):
@@ -339,7 +341,7 @@ class APITest(unittest.TestCase):
                 np.testing.assert_almost_equal(
                     series.get_attribute("nparr_cdouble"),
                     [4.5 + 1.1j, 6.7 - 2.2j])
-            if file_ending != "bp":  # not in ADIOS 1.13.1 nor ADIOS 2.6.0
+            if file_ending != "bp":  # not in ADIOS 1.13.1 nor ADIOS 2.7.0
                 np.testing.assert_almost_equal(
                     series.get_attribute("nparr_clongdouble"),
                     [8.9 + 7.8j, 7.6 + 9.2j])